requirements.txt
python-dotenv==1.0.0
redis==5.0.0